refactor: remove unused constants and move non-shared constants to respective files
