+++ src/rds-control-plane-mcp-server/awslabs/rds_control_plane_mcp_server/resources/db_instance/read_performance_report.py
@@ -16,7 +16,7 @@
 
 import asyncio
 from ...common.connection import PIConnectionManager
-from ...common.exceptions import handle_exceptions
+from ...common.decorators.handle_exceptions import handle_exceptions
 from ...common.server import mcp
 from datetime import datetime
 from loguru import logger
